back-end/requirements.txt
alembic==1.13.3
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.0
bleach==6.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
pycparser==2.22
pytz==2024.2
setuptools==75.3.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.0.6
WTForms==3.2.1
zxcvbn==4.4.28
PyGithub==2.5.0